Release version 0.1.0
This is an initial release. Nothing much has changed.

Remember to create a release on GitHub. First, push all your
changes. Next, go to your GitHub repository page, and click on the
button saying 0 releases, and next on create a new release. You are
asked to provide three things

- The tag version should be v0.1.0.
- The Release title can be 'Initial release'.
- What to set as the description for this initial release is up to you, it can for example be one or two sentences about your project.

It may be a good idea to mark your project as a pre-release until your
are confident that it is widely usable. Finally, click Public
release. GitHub will now create a new tag (but not a new commit)
permanently marking this version of your code as v0.1.0. Optionally,
you can immediately pull this tag to your local repository:

$ git pull --tags

For more information, check out

https://cicwi.github.io/software-guides/software-release-guide