Add Astra-Toolbox as a dependency
Adding astra as a dependency requires changes to multiple files.

setup.py
--------

Here you can must astra-toolbox to the requirements array. In
practice, you will almost always want to add numpy as well. I have not
put any version constraints on numpy, since astra will usually require
a rather specific version of numpy anyhow.

Makefile
--------

To build the package, `conda build` has to be able to find the
`astra-toolbox` package. Astra is not part of the default conda
distribution, but part of a separate 'channel'.  Therefore, you must
add the astra channel by appending `-c astra-toolbox` or `-c
astra-toolbox/label/dev` to the command in the Makefile. The latter of
which provides the development channel of Astra.

If you do not update the Makefile, you will get errors like:

```
conda.exceptions.ResolvePackageNotFound:
  - astra-toolbox
...
conda_build.exceptions.DependencyNeedsBuildingError: Unsatisfiable dependencies for platform linux-64: {'astra-toolbox'}

During handling of the above exception, another exception occurred:
```

README.md
---------

Of course, do not forget to update the installation instructions!

You can build the conda package using

$ cd <to-your-project-directory>
$ make conda_package

NOTE: you have not yet bumped the version of your package, so it is
unwise to upload this package to the anaconda cloud. We will bump the
version in the next few commits.