Bump version 0.1.0 -> 0.2.0
Bumping the version
-------------------

Bumping a version involves updating the version number in three
positions:
- setup.py
- cookiecutter_example/__init__.py
- setup.cfg

You can do this manually, or you can automate this. If you have the
development dependencies installed you can automate this process by
executing:

$ bumpversion --no-tag --no-commit minor

This will update the `minor` version number (i.e., the middle
one). The other options are `major` and `patch` for the first and last
number respectively.

NOTE: it is highly recommonded to use the flags `--no-tag` and
`--no-commit`. Otherwise, `bumpversion` will automatically create git
commits and tags, which is terrible if you also want to update the
changelog.

If bumpversion is not yet installed, execute:
```
make install_dev
```

Updating changelog
------------------

Of course, a new release means that you must update CHANGELOG.md. Here
you can create a new section for the current release and create a new
clean section for unreleased code. As a courtesy to your users, you
will want to add the link to compare the code in the current version
to the code in the previous version. This is done in the last line.

Documentation
-------------

Normally, you will want to have up to date documentation as well. In
this commit, it is not necessary to update the documentation since it
has been generated in the previous commit.

Further actions
---------------

There are two more things to do:
1) Create a GitHub release
2) Upload a new Conda package

Creating a GitHub release is explained in one of the previous
commits. See the common actions index.

To upload a new conda package, execute:
$ make conda_package
$ anaconda upload <the-path-to-the-package-that-is-displayed-in-the-output>

This step has also been done before. If anything is unclear, be sure
to look it up in the common actions index!